+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # Changelog
 
+## [4.2.0](https://github.com/dunay2/dvt/compare/v4.1.2...v4.2.0) (2026-03-16)
+
+
+### Features
+
+* Add G7 projector worker runtime ([#487](https://github.com/dunay2/dvt/issues/487)) ([8c6f31e](https://github.com/dunay2/dvt/commit/8c6f31e276dc08f242faa3193a158f9cd4863df0))
+
 ## [4.1.2](https://github.com/dunay2/dvt/compare/v4.1.1...v4.1.2) (2026-03-16)
 
 
 ### Bug Fixes
 
 * **adapters:** Harden outbox-worker cleanup error serialization ([#483](https://github.com/dunay2/dvt/issues/483)) ([d97f755](https://github.com/dunay2/dvt/commit/d97f755f08568fca0b469f9dcef0202b998d7a2a))
+* **ci:** Harden workspace prebuild chains ([#485](https://github.com/dunay2/dvt/issues/485)) ([91fc13a](https://github.com/dunay2/dvt/commit/91fc13ac2793257d33bdca8574e556fdebb1bbc0))
 
 ## [4.1.1](https://github.com/dunay2/dvt/compare/v4.1.0...v4.1.1) (2026-03-16)